generater should return float not int

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -8,7 +8,7 @@ import asyncio
 import random
 from typing import Generator
 
-async def async_generator() -> Generator[int, None, None]:
+async def async_generator() -> Generator[float, None, None]:
     """generates a random float number in the given range"""
     for i in range(0, 10, 1):
         await asyncio.sleep(1)
